requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
django-soft-delete==1.0.15
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.7
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
